Working particle generator.
