+++ packages/particles-ui-demo/src/prompts/si.ts
@@ -1,0 +1,8 @@
+import preamble from "./si-preamble.md?raw";
+import toolkitTypes from "../types/particles?raw";
+
+export default `
+${preamble}
+
+${toolkitTypes}
+`;
